# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST API"
  },
  "paths": {
...
    "/api/conversion/html-to-pdf": {
      "post": {
        "operationId": "convertHtmlToPdf",
        "parameters": [
          {
            "description": "default value: landscape",
            "in": "query",
            "name": "orientation",
            "schema": {
              "description": "Enum representing print orientations",
              "enum": [
                "PORTRAIT",
                "LANDSCAPE"
              ],
              "type": "string"
            }
          },
          {
            "description": "default value: A4",
            "in": "query",
            "name": "paperSize",
            "schema": {
              "description": "Enum representing print paper sizes",
              "enum": [
                "A4",
                "A3"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/html": {
              "schema": {
                "type": "string"
              }
            }
          },
          "description": "input html (must include html and body elements)"
        },
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "Content of PDF document as a byte array"
          }
        },
        "summary": "Converts input HTML to PDF",
        "tags": [
          "Conversion"
        ]
      }
    },
    "/api/diff/documents": {
      "post": {
        "operationId": "getDocumentsDiff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentsDiffParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsDiff"
                }
              }
            },
            "description": "Successfully retrieved the differences between the provided documents"
          }
        },
        "summary": "Gets difference of two live documents",
        "tags": [
          "Difference"
        ]
      }
    },
    "/api/diff/html": {
      "post": {
        "operationId": "diffHtml",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "html1": {
                    "type": "string"
                  },
                  "html2": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringsDiff"
                }
              }
            },
            "description": "Successfully retrieved the differences between the provided HTML strings"
          }
        },
        "summary": "Gets difference of two strings which contain HTML tags",
        "tags": [
          "Difference"
        ]
      }
    },
    "/api/diff/text": {
      "post": {
        "operationId": "diffText",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "text1": {
                    "type": "string"
                  },
                  "text2": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringsDiff"
                }
              }
            },
            "description": "Successfully retrieved the differences between the provided plain text strings"
          }
        },
        "summary": "Gets difference of two strings which contain plain text",
        "tags": [
          "Difference"
        ]
      }
    },
    "/api/diff/workitems": {
      "post": {
        "operationId": "getWorkItemsDiff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkItemsDiffParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemsDiff"
                }
              }
            },
            "description": "Successfully retrieved the differences between the provided WorkItems"
          }
        },
        "summary": "Gets difference of two WorkItems",
        "tags": [
          "Difference"
        ]
      }
    },
    "/api/merge/workitems": {
      "post": {
        "operationId": "mergeWorkItems",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeResult"
                }
              }
            },
            "description": "Merge result"
          }
        },
        "summary": "Merge WorkItems",
        "tags": [
          "Merge"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Document": {
        "description": "Represents a document within a project, including metadata such as projectID, space ID, and revision details",
        "properties": {
          "authorizedForMerge": {
            "description": "Indicates whether the user is authorized for merge operations",
            "type": "boolean"
          },
          "headRevision": {
            "description": "The head revision of the document",
            "type": "string"
          },
          "id": {
            "description": "The unique identifier of the document",
            "type": "string"
          },
          "locationPath": {
            "description": "The location path of the document",
            "type": "string"
          },
          "moduleXmlRevision": {
            "description": "The XML revision of the module",
            "type": "string"
          },
          "projectId": {
            "description": "The ID of the project this document belongs to",
            "type": "string"
          },
          "projectName": {
            "description": "The name of the project this document belongs to",
            "type": "string"
          },
          "revision": {
            "description": "The current revision of the document",
            "type": "string"
          },
          "spaceId": {
            "description": "The ID of the space where the document is located",
            "type": "string"
          },
          "spaceName": {
            "type": "string"
          },
          "title": {
            "description": "The title of the document",
            "type": "string"
          }
        },
        "type": "object"
      },
      "DocumentIdentifier": {
        "description": "Unique document identifier data",
        "properties": {
          "moduleXmlRevision": {
            "description": "XML module revision number",
            "type": "string"
          },
          "name": {
            "description": "Name of the document",
            "type": "string"
          },
          "projectId": {
            "description": "ID of the project",
            "type": "string"
          },
          "revision": {
            "description": "Revision number of the document",
            "type": "string"
          },
          "spaceId": {
            "description": "ID of the space",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DocumentWorkItem": {
        "description": "WorkItem in document data",
        "properties": {
          "documentLocationPath": {
            "description": "The location path of the document containing the WorkItem",
            "type": "string"
          },
          "documentProjectId": {
            "description": "The project ID of the document containing the WorkItem",
            "type": "string"
          },
          "documentRevision": {
            "description": "The revision of the document containing the WorkItem",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier of the WorkItem",
            "type": "string"
          },
          "projectId": {
            "description": "WorkItem's project ID, it can differ from document's project ID in case of referenced WorkItems",
            "type": "string"
          },
          "revision": {
            "description": "The revision of the WorkItem. Referenced WorkItem's can be frozen in document in certain revision",
            "type": "string"
          }
        },
        "type": "object"
      },
      "DocumentsDiff": {
        "description": "Represents the difference between two documents",
        "properties": {
          "extensionInfo": {
            "description": "Information about extension (version, build timestamp)",
            "type": "string"
          },
          "leftDocument": {
            "$ref": "#/components/schemas/Document"
          },
          "pairedWorkItems": {
            "$ref": "#/components/schemas/WorkItemsPair"
          },
          "rightDocument": {
            "$ref": "#/components/schemas/Document"
          }
        },
        "type": "object"
      },
      "DocumentsDiffParams": {
        "description": "Comparison input data",
        "properties": {
          "configCacheBucketId": {
            "description": "The ID of the configuration cache bucket",
            "type": "string"
          },
          "configName": {
            "description": "The configuration name to use for the comparison",
            "type": "string"
          },
          "leftDocument": {
            "$ref": "#/components/schemas/DocumentIdentifier"
          },
          "linkRole": {
            "description": "The role of the link connecting the two documents",
            "type": "string"
          },
          "rightDocument": {
            "$ref": "#/components/schemas/DocumentIdentifier"
          }
        },
        "type": "object"
      },
      "Field": {
        "description": "Represents a field within a work item",
        "properties": {
          "html": {
            "description": "HTML representation of the field",
            "type": "string"
          },
          "htmlDiff": {
            "description": "HTML diff representation of the field",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier of the field",
            "type": "string"
          },
          "issues": {
            "description": "Set of issues related to the field",
            "items": {
              "description": "Set of issues related to the field",
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "name": {
            "description": "Name of the field",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/IType"
          },
          "value": {
            "description": "Value of the field",
            "type": "object"
          }
        },
        "type": "object"
      },
      "FieldDiff": {
        "description": "Represents the differences between fields in WorkItems",
        "properties": {
          "diffLeft": {
            "description": "Difference on the left side",
            "type": "string"
          },
          "diffRight": {
            "description": "Difference on the right side",
            "type": "string"
          },
          "id": {
            "description": "Field ID",
            "type": "string"
          },
          "issues": {
            "description": "Set of issues related to this field",
            "items": {
              "description": "Set of issues related to this field",
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "name": {
            "description": "Field name",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "IType": {
        "description": "Type of the field",
        "type": "object"
      },
      "MergeParams": {
        "description": "Merge input data",
        "properties": {
          "configCacheBucketId": {
            "description": "The ID of the configuration cache bucket",
            "type": "string"
          },
          "configName": {
            "description": "The configuration name to use for the merge operation",
            "type": "string"
          },
          "direction": {
            "description": "Enum representing the direction of a merge operation",
            "enum": [
              "LEFT_TO_RIGHT",
              "RIGHT_TO_LEFT"
            ],
            "type": "string"
          },
          "leftDocument": {
            "$ref": "#/components/schemas/DocumentIdentifier"
          },
          "linkRole": {
            "description": "The role of the link connecting the two documents",
            "type": "string"
          },
          "pairs": {
            "$ref": "#/components/schemas/WorkItemsPair"
          },
          "rightDocument": {
            "$ref": "#/components/schemas/DocumentIdentifier"
          }
        },
        "type": "object"
      },
      "MergeResult": {
        "description": "Result of a merge operation",
        "properties": {
          "conflictedPairs": {
            "$ref": "#/components/schemas/WorkItemsPair"
          },
          "createdPairs": {
            "$ref": "#/components/schemas/WorkItemsPair"
          },
          "mergeNotAuthorized": {
            "description": "Indicates whether the merge was not authorized",
            "type": "boolean"
          },
          "modifiedPairs": {
            "$ref": "#/components/schemas/WorkItemsPair"
          },
          "movedPairs": {
            "$ref": "#/components/schemas/WorkItemsPair"
          },
          "notMovedPairs": {
            "$ref": "#/components/schemas/WorkItemsPair"
          },
          "success": {
            "description": "Indicates whether the merge operation was successful",
            "type": "boolean"
          },
          "targetModuleHasStructuralChanges": {
            "description": "Indicates whether the target module has structural changes",
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "StringsDiff": {
        "description": "Represents the result of comparing two strings",
        "properties": {
          "different": {
            "description": "Indicates whether the two strings are different",
            "type": "boolean"
          },
          "result": {
            "description": "The result after comparing the strings",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "WorkItem": {
        "description": "Represents a WorkItem",
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/Field"
            },
            "type": "array"
          },
          "id": {
            "description": "Unique identifier of the WorkItem",
            "type": "string"
          },
          "lastRevision": {
            "description": "Last revision number of the WorkItem",
            "type": "string"
          },
          "moveDirection": {
            "description": "Enum representing the direction in which a merge move operation can occur",
            "enum": [
              "UP",
              "DOWN"
            ],
            "type": "string"
          },
          "movedOutlineNumber": {
            "description": "Moved outline number of the work item after a merge",
            "type": "string"
          },
          "outlineNumber": {
            "description": "Outline number of the WorkItem",
            "type": "string"
          },
          "projectId": {
            "description": "Project ID associated with the WorkItem",
            "type": "string"
          },
          "referenced": {
            "description": "Flag indicating whether the WorkItem is referenced",
            "type": "boolean"
          },
          "revision": {
            "description": "Revision number of the WorkItem",
            "type": "string"
          },
          "title": {
            "description": "Title of the WorkItem",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "WorkItemsDiff": {
        "description": "Represents the differences between two WorkItems",
        "properties": {
          "fieldDiffs": {
            "items": {
              "$ref": "#/components/schemas/FieldDiff"
            },
            "type": "array"
          },
          "leftWorkItem": {
            "$ref": "#/components/schemas/WorkItem"
          },
          "rightWorkItem": {
            "$ref": "#/components/schemas/WorkItem"
          }
        },
        "type": "object"
      },
      "WorkItemsDiffParams": {
        "description": "Comparison input data",
        "properties": {
          "configCacheBucketId": {
            "description": "Cache bucket Id",
            "type": "string"
          },
          "configName": {
            "description": "Name of the configuration used for comparison",
            "type": "string"
          },
          "leftProjectId": {
            "description": "ID of the left project",
            "type": "string"
          },
          "leftWorkItem": {
            "$ref": "#/components/schemas/DocumentWorkItem"
          },
          "pairedWorkItemsDiffer": {
            "description": "Indicates if the paired WorkItems differ",
            "type": "boolean"
          },
          "pairedWorkItemsLinkRole": {
            "description": "Link role between paired WorkItems",
            "type": "string"
          },
          "rightWorkItem": {
            "$ref": "#/components/schemas/DocumentWorkItem"
          }
        },
        "type": "object"
      },
      "WorkItemsPair": {
        "description": "Represents a pair of WorkItems for comparison",
        "properties": {
          "leftWorkItem": {
            "$ref": "#/components/schemas/WorkItem"
          },
          "rightWorkItem": {
            "$ref": "#/components/schemas/WorkItem"
          }
        },
        "type": "object"
      }
    }
  }
}